Initial Commit
This includes the basic building blocks for the provee library as
outlined in the README.md. It includes a few ensure modules for managing
files, packages, services and arbitrary command executions. It includes
an example project that should be used as a getting started guide for new
users.

ffi.py is a small script that was used to discover the structure,
including the primitive types, of various structs that were needed to
eventually implement simple procedures for changing/retrieving the
owner/group of files. Unfortunately, these procedures aren't included in
the Racket standard library.